# Conflicts:
#	tt-inference-server-v2/llm_module/config.py
#	tt-inference-server-v2/test_module/llm_tests/agentic_eval_tests.py

+++ tt-media-server/cpp_server/src/sockets/inter_server_service.cpp
@@ -133,7 +133,7 @@ bool InterServerService::isEnabled() const { return enabled; }
 
 bool InterServerService::sendPrefillRequest(
     uint32_t taskId, const std::vector<uint64_t>& registrationHashes,
-    const std::vector<int64_t>& tokenIds, std::optional<int> maxTokens,
+    const std::vector<uint32_t>& tokenIds, std::optional<int> maxTokens,
     std::optional<uint32_t> slotId,
     const tt::domain::llm::SamplingParams& sampling, int decodePositionId,
     int decodeSkipTokens) {
